requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.1
requests==2.31.0
# matplotlib==3.7.2
# celery==5.3.1
# django-celery-beat==2.5.0
# django-celery-results==2.5.1
# redis==4.6.0
dj-database-url==2.0.0
